refactor(ui): move reply-preview resolution into its own module
chat-transcript-projection crossed the 700-line cap after rebasing onto
main; the reply-preview concern (loaded-source index type, preview
projection, memoized resolver) moves to chat-reply-preview.ts.
